% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{get_control}
\alias{get_control}
\title{Gibbs control object}
\usage{
get_control(R, burn = 0, thin = 1, report_period = R + 1, save_latent = FALSE)
}
\arguments{
\item{R}{Desired number of draws from the sampler.}

\item{burn}{Number of draws to burn before saving.}

\item{thin}{The period for saving samples after burn phase has ended.}

\item{report_period}{The period for printing progress.}

\item{save_latent}{Logical; if true, save draws of the latent variable
\eqn{mu(\bm{X})}.}
}
\value{
An object of class \code{my_control}.
}
\description{
Gibbs control object
}
