% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed.R
\name{get_fixed}
\alias{get_fixed}
\title{Gibbs fixed values object}
\usage{
get_fixed(muX = FALSE, sigma2 = FALSE)
}
\arguments{
\item{muX}{Logical}

\item{sigma2}{Logical}
}
\value{
An object of class \code{my_fixed}.
}
\description{
Specify conditionals which should be kept fixed (i.e., not redrawn) during
sampling.
}
