% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{get_prior}
\alias{get_prior}
\title{Gibbs fixed values object}
\usage{
get_prior(a_sigma, b_sigma)
}
\arguments{
\item{a_sigma}{A positive number}

\item{b_sigma}{A positive number}
}
\value{
An object of class \code{my_prior}.
}
\description{
Specify hyperparameters \eqn{a_\sigma} and \eqn{b_\sigma} for Inverse Gamma
prior on \eqn{\sigma^2}.
}
