% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{gibbs}
\alias{gibbs}
\alias{summary.gputil_fit}
\alias{print.gputil_fit}
\alias{predict.gputil_fit}
\title{Gibbs sampler for Gaussian Process regression model}
\usage{
gibbs(
  y,
  K,
  init = get_init(),
  prior = get_prior(),
  control = get_control(),
  fixed = get_fixed()
)

\method{summary}{gputil_fit}(object, ...)

\method{print}{gputil_fit}(x, ...)

\method{predict}{gputil_fit}(object, K22, K21)
}
\arguments{
\item{y}{Vector of observations}

\item{K}{The matrix \eqn{k(X, X)}; see details.}

\item{init}{Initial values object from \code{get_init}}

\item{prior}{Hyperparameters object from \code{get_prior}}

\item{control}{Control object from \code{get_control}}

\item{fixed}{Object declaring fixed quantities from \code{get_fixed}}

\item{object}{An object of class \code{gputil_fit} returned by the Gibbs
sampler.}

\item{x}{An object of class \code{gputil_fit} returned by the Gibbs
sampler.}

\item{K22}{The matrix \eqn{k(X_0, X_0)}; see details.}

\item{K21}{The matrix \eqn{k(X_0, X)}; see details.}
}
\value{
The \code{gibbs} function returns an object with sampler results. The
\code{predict} function returns an \eqn{R \times n_0} matrix of draws from
the posterior predictive distribution.
}
\description{
Gibbs sampler for Gaussian Process regression model
}
\details{
This is a Gibbs sampler for the model
\deqn{
\bm{y} = \mu(\bm{X}) + \bm{\epsilon}, \quad \bm{\epsilon} \sim
\text{N}(\bm{0}, \sigma^2 \bm{I}), \\
\mu(\bm{X}) \sim \text{N}(\bm{0}, k(\bm{X},\bm{X})) \\
\sigma^2 \sim \text{IG}(a_\sigma, b_\sigma)
}

We have used the notation \eqn{\bm{X}} for the \eqn{n \times d} matrix of
\eqn{n} inputs whose corresponding observations are
\eqn{\bm{y} = (y_1, \ldots, y_n)}. We will also write \eqn{\bm{X}_0} as an
\eqn{n_0 \times d} matrix of inputs
\eqn{\bm{x}_{01}, \ldots, \bm{x}_{0 n_0}} on which to compute predictions.
The user may choose any covariance kernel \eqn{k(\bm{x}, \bm{x}')}, but it
is their responsibility to evaluate it on the inputs. Any parameters of the
kernel are considered fixed during MCMC.

The \code{gibbs} functons requires the matrix \eqn{k(\bm{X}, \bm{X})} whose
\eqn{(i,j)}th element is \eqn{k(\bm{x}_i, \bm{x}_j)}. This is the argument
\code{K}. \code{K} may be provided as a matrix or the result of a spectral
decomposition; if \code{K} is a \code{list}, we assume eigenvalues are in
\code{K$values} and \code{K$vectors} is a matrix whose columns are the
corresponding eigenvectors. Otherwise, if \code{K} is not a \code{list}, it
is interpreted as a matrix.

The \code{predict} function requires the \eqn{n_0 \times n_0} matrix
\eqn{k(\bm{X}_0, \bm{X}_0) = k(\bm{x}_{0i}, \bm{x}_{0j})} as argument
\code{K22}, and the matrix \eqn{k(\bm{X}_0, \bm{X}) = k(\bm{x}_{0i}, \bm{x}_j)}
as argument \code{K21}.
}
